+++ app/players/helpers.py
@@ -162,7 +162,11 @@ def get_hero_role(hero_key: HeroKey) -> Role | None:  # ty: ignore[invalid-type-
     """Get the role of a given hero based on the CSV file"""
     heroes_data = read_csv_data_file("heroes")
     role_key = next(
-        (hero_data["role"] for hero_data in heroes_data if hero_data["key"] == hero_key),
+        (
+            hero_data["role"]
+            for hero_data in heroes_data
+            if hero_data["key"] == hero_key
+        ),
         None,
     )
     if role_key is None: